extract Tooltip out from TenantProfile
