+++ apps/dashboard/components/PropertyDetails/TenantProfile.tsx
@@ -1,11 +1,9 @@
 import {Avatar} from '@/volto/Avatar';
 import {Card} from '@/volto/Card';
 import {Check} from '@/volto/icons';
-import {useTooltipsManager} from '@/volto/Tooltip';
-import {AriaPositionProps, useOverlayPosition} from '@react-aria/overlays';
+import {Tooltip, useTooltip} from '@/volto/Tooltip';
 import Link from 'next/link';
-import {PropsWithChildren, useEffect, useId, useRef, useState} from 'react';
-import {createPortal} from 'react-dom';
+import {useEffect, useRef, useState} from 'react';
 import * as s from './TenantProfile.css';
 
 type TenantProfileProps = {
@@ -45,8 +43,7 @@ export function TenantProfile({
 }
 
 function Contact({children}: {children: string}) {
-  const tooltipId = useId();
-  const tooltipState = useTooltipState(tooltipId);
+  const tooltip = useTooltip();
   const tooltipTriggerRef = useRef<HTMLButtonElement>(null);
 
   const [isCopied, setCopied] = useState(false);
@@ -59,27 +56,27 @@ function Contact({children}: {children: string}) {
   }
 
   function handleTooltipClose() {
-    tooltipState.close();
+    tooltip.state.close();
     setCopied(false);
   }
 
   return (
     <>
       <button
-        aria-describedby={tooltipState.isOpen ? tooltipId : undefined}
+        aria-describedby={tooltip.state.isOpen ? tooltip.id : undefined}
         className={s.Contact}
         onClick={handleClick}
-        onFocus={tooltipState.open}
+        onFocus={tooltip.state.open}
         onBlur={handleTooltipClose}
-        onMouseEnter={tooltipState.open}
+        onMouseEnter={tooltip.state.open}
         onMouseLeave={handleTooltipClose}
         ref={tooltipTriggerRef}
       >
         {children}
       </button>
       <Tooltip
-        id={tooltipId}
-        isOpen={tooltipState.isOpen}
+        id={tooltip.id}
+        isOpen={tooltip.state.isOpen}
         placement="bottom"
         targetRef={tooltipTriggerRef}
       >
@@ -120,86 +117,4 @@ function CopiedContact({
       Copied
     </div>
   );
-}
-
-// Tooltip is an element to provide a short context or hint about an UI element.
-
-type TooltipProps = PropsWithChildren<{
-  id: string;
-  isOpen: boolean;
-}> &
-  Pick<AriaPositionProps, 'placement' | 'targetRef'>;
-
-// Tooltip.Overlay
-function Tooltip(props: TooltipProps) {
-  const {children, id, isOpen, placement, targetRef} = props;
-
-  const {portalsContainer} = useTooltipsManager();
-  const overlayRef = useRef<HTMLDivElement>(null);
-  const {overlayProps, updatePosition} = useOverlayPosition({
-    boundaryElement: portalsContainer || undefined,
-    isOpen,
-    overlayRef,
-    placement,
-    targetRef,
-  });
-
-  // update position when content (children) changes
-  useEffect(() => {
-    if (isOpen) {
-      updatePosition();
-    }
-  }, [children, isOpen, updatePosition]);
-
-  if (!isOpen || !portalsContainer) {
-    return null;
-  }
-
-  const position = overlayProps.style ?? {};
-  return createPortal(
-    <div
-      id={id}
-      className={s.TooltipContainer}
-      ref={overlayRef}
-      role="tooltip"
-      style={{
-        top: position.top,
-        left: position.left,
-        bottom: position.bottom,
-        right: position.right,
-      }}
-    >
-      {children}
-    </div>,
-    portalsContainer
-  );
-}
-
-function useTooltipState(id: string, initialOpen: boolean = false) {
-  const [isOpen, setOpen] = useState(initialOpen);
-  const {visibility} = useTooltipsManager();
-
-  useEffect(
-    () => () => {
-      visibility.removeCloseFunction(id);
-    },
-    [id, visibility]
-  );
-
-  return {
-    isOpen,
-
-    open() {
-      visibility.closeAll();
-      visibility.addCloseFunction(id, () => setOpen(false));
-      setOpen(true);
-    },
-
-    close() {
-      // need to unsubscribe on close because closing a tooltip
-      // does not guarantee to cause an unmount.
-      visibility.removeCloseFunction(id);
-      setOpen(false);
-    },
-  };
 }
