+++ apps/dashboard/volto/Tooltip/TooltipsManager.css.ts
@@ -1,7 +1,5 @@
 import {style} from '@vanilla-extract/css';
 
-// export const Tooltip = style({});
-
 export const PortalsContainer = style({
   position: 'relative',
 });
