+++ apps/dashboard/volto/Tooltip/index.ts
@@ -1,0 +1,3 @@
+export {Tooltip} from './Tooltip';
+export {TooltipsManagerProvider} from './TooltipsManager';
+export {useTooltip} from './useTooltip';
